Refactor ChannelI18nUtil to simple utility class
